+++ src/main/scala/chisel3/tester/exceptions.scala
@@ -1,0 +1,9 @@
+// See LICENSE for license details.
+
+package chisel3.tester
+
+class ThreadOrderDependentException(message: String) extends Exception(message)
+class TimeoutException(message: String) extends Exception(message)
+
+// when interfacing with the testdriver before stepping the clock after moving to an earlier region
+class TemporalParadox(message: String) extends Exception(message)
