+++ src/main/scala/chisel3/tester/experimental/package.scala
@@ -1,0 +1,11 @@
+// See LICENSE for license details.
+
+package chisel3.tester
+
+import chisel3._
+
+package object experimental {
+  implicit class uncheckedPokeableClock(x: Clock) {
+
+  }
+}
