+++ src/main/scala/chisel3/tester/internal/TestEnvInterface.scala
@@ -1,6 +1,6 @@
 // See LICENSE for license details.
 
-package chisel3.tester
+package chisel3.tester.internal
 
 import scala.collection.mutable
