+++ src/main/scala/chisel3/tester/internal/Testers2.scala
@@ -1,13 +1,14 @@
 // See LICENSE for license details.
 
-package chisel3.tester
+package chisel3.tester.internal
 
 import scala.util.DynamicVariable
 import chisel3.experimental.MultiIOModule
-import firrtl.ExecutionOptionsManager
+
+import scala.util.DynamicVariable
 
 // Internal common set of options to be understood by all backends
-private[tester] case class TesterOptions(
+case class TesterOptions(
   name: String,
   writeVcd: Boolean
 )
@@ -23,14 +24,6 @@ object Context {
     context.withValue(Some(new Instance(backend, env))) {
       backend.run(testFn)
     }
-  }
-
-  // TODO: better integration points for default tester selection
-  // TODO: add TesterOptions (from chisel-testers) and use that to control default tester selection.
-  def createDefaultTester[T <: MultiIOModule](dutGen: () => T, testOptions: TesterOptions,
-      execOptions: Option[ExecutionOptionsManager]
-      ): BackendInstance[T] = {
-    TreadleExecutive.start(dutGen, testOptions, execOptions)
   }
 
   def apply(): Instance = context.value.get
