+++ src/main/scala/chisel3/tester/package.scala
@@ -3,8 +3,8 @@
 package chisel3
 
 import scala.language.implicitConversions
-
-import chisel3.core.ActualDirection  // TODO needs to be a public API
+import chisel3.tester.internal._
+import chisel3.core.ActualDirection
 import chisel3.experimental.{DataMirror, FixedPoint}
 import chisel3.util._
 
@@ -122,21 +122,6 @@ package object tester {
   }
 
   object fork extends ForkBuilder(None, None, Seq())
-
-  case class ForkBuilder(name: Option[String], region: Option[Region], threads: Seq[AbstractTesterThread]) {
-    def apply(runnable: => Unit): TesterThreadList = {
-      new TesterThreadList(threads ++ Seq(Context().backend.doFork(() => runnable, name, region)))
-    }
-
-    def withRegion(newRegion: Region): ForkBuilder = {
-      require(region.isEmpty)
-      this.copy(region=Some(newRegion))
-    }
-    def withName(newName: String): ForkBuilder = {
-      require(name.isEmpty)
-      this.copy(name=Some(newName))
-    }
-  }
 
   // TODO: call-by-name doesn't work with varargs, is there a better way to do this?
   def parallel(run1: => Unit, run2: => Unit): Unit = {
